retrieving data from the API, store it in room in one mode

+++ app/src/main/java/dog/snow/androidrecruittest/data/model/VPRawPhoto.kt
@@ -5,8 +5,8 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class VPRawPhoto(
-    val id: Int,
-    val albumId: Int,
+    val id: Long,
+    val albumId: Long,
     val title: String,
     val url: String,
     val thumbnailUrl: String
